Including Java 21 in Github Actions runs.
Upgraded Gradle to 8.5

Fixed a few new linting problems, caught by '-Werror'.

Had to turn off linting "this-escape" as the Wrapper thingy violates
this, and that is a part of the API.

Moved a few test classes to get more even run.

+++ mats-impl-jms/src/main/java/io/mats3/impl/jms/JmsMatsStartStoppable.java
@@ -21,7 +21,7 @@ public interface JmsMatsStartStoppable extends JmsMatsStatics, StartStoppable {
         for (JmsMatsStartStoppable child : getChildrenStartStoppable()) {
             long millisBefore = System.currentTimeMillis();
             started &= child.waitForReceiving(millisLeft);
-            millisLeft -= (System.currentTimeMillis() - millisBefore);
+            millisLeft -= (int) (System.currentTimeMillis() - millisBefore);
             millisLeft = Math.max(millisLeft, EXTRA_GRACE_MILLIS);
         }
         return started;
@@ -42,7 +42,7 @@ public interface JmsMatsStartStoppable extends JmsMatsStatics, StartStoppable {
         for (JmsMatsStartStoppable child : getChildrenStartStoppable()) {
             long millisBefore = System.currentTimeMillis();
             child.stopPhase2_GracefulWaitAfterRunflagFalse(millisLeft);
-            millisLeft -= (System.currentTimeMillis() - millisBefore);
+            millisLeft -= (int) (System.currentTimeMillis() - millisBefore);
             millisLeft = Math.max(millisLeft, EXTRA_GRACE_MILLIS);
         }
     }
